backend/requirements.txt
# =================================
# CORE FRAMEWORK
# =================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# =================================
# DATABASE
# =================================
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1

# =================================
# AI & ML
# =================================
openai==1.3.7
anthropic==0.7.1
langchain==0.1.0
langchain-openai==0.0.2

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Video processing
moviepy==1.0.3
ffmpeg-python==0.2.0

# =================================
# SOCIAL MEDIA APIS
# =================================
# Meta/Facebook
facebook-business==18.0.2

# Google
google-ads==23.0.0
google-api-python-client==2.108.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0

# TikTok (using requests for custom API calls)
# YouTube
google-api-python-client==2.108.0

# =================================
# TASK QUEUE
# =================================
celery[redis]==5.3.4
flower==2.0.1

# =================================
# HTTP & NETWORKING
# =================================
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# =================================
# AUTHENTICATION & SECURITY
# =================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1

# =================================
# STORAGE
# =================================
boto3==1.34.0  # AWS S3
botocore==1.34.0

# =================================
# UTILITIES
# =================================
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0

# File handling
python-magic==0.4.27

# =================================
# MONITORING & LOGGING
# =================================
sentry-sdk[fastapi]==1.38.0
loguru==0.7.2

# =================================
# DEVELOPMENT & TESTING
# =================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
isort==5.13.0
flake8==6.1.0
mypy==1.7.1

# =================================
# DATA VALIDATION
# =================================
email-validator==2.1.0
phonenumbers==8.13.26

# =================================
# RATE LIMITING
# =================================
slowapi==0.1.9

# =================================
# CORS
# =================================
# Included in FastAPI

# =================================
# BACKGROUND JOBS
# =================================
apscheduler==3.10.4

# =================================
# TEXT PROCESSING
# =================================
langdetect==1.0.9
deep-translator==1.11.4
